Refactor Jekyll::ShowDataGenerator, re #194
Not 100% on the functionality of this, mainly because of the mess
it was before!

Could probably be cleaned up further, don't really like how
processing happens on the show object in one of two iterations:
pre-sort and post-sort.

I've made the decision to keep most of the functions in pre-sort
and only include what needs to be in post-sort in post-sort.