TJP: Add config to use vagrant for testing.
Got fed up working with NP to get it running on Windows and throught
an easy way to run a VM would be the more sensbile suggestion.

Pretty much there, just need to find a way of running the build script
and python webserver in a nice way. This should be fairly trivial to do.