Move JS to <head> from end of <body>
When in body, the library and app JS has to run on every navigation.
Even though we're running turbolinks. This is silly. Now JS in in head
it is evaluated and run once when the page is loaded (and still when
we search).

This breaks event listeners for document.ready. They will be fixed in
the next commit.