Switch to showing focus state using focus-ring
The pseudo-selector :focus-ring allows for showing a 'focus-ring' when
navigating with the keyboard but without showing it for mouse users.
This means we can make the focus-ring obnoxiously obvious without
sacrificing aesthetics for mouse users.

As :focus-ring [1] is very new (so new it's not even on caniuse.com
yet—we be hipster), we needs a polyfill [2]. Add this.

See Google's a11y cast video for implementation description [3].

[1]: https://drafts.csswg.org/selectors-4/#the-focusring-pseudo
[2]: https://github.com/WICG/focus-ring
[3]: https://www.youtube.com/watch?v=ilj2P5-5CjI&list=PLNYkxOF6rcICWx0C9LVWWVqvHlYJyqw7g&index=1